refactor: custom hook으로 로직 분리
